removed some unused ut and add comments
